Make manage.py be able to find the default fixtures yaml even if it's not run from the trigger_webapp subdirectory
Make migrations first?

Is not being in a the right directory causing problems?

Change database engine

Removed the default fixtures as that was causing problems

Do a normal migration and a syncdb migration

Specified the app in the migrations

Run test in trigger-webapp directory so it can find the fixtures

Still need a mwa key even if it won't work